+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Improved clipboard persistence: Vendace will now stay alive until the copied URL is overwritten.
 
+* Added a setting for copying to the selection buffer.
+
 ## Version 0.1.0 (June 3rd, 2014)
 
 **First release**. Supports screen capture, simple cropping, external editing and upload.